refactor: API 메서드 네이밍 변경
